requirements.txt
# AI Smart Car Project Dependencies

# Web Framework
flask==2.3.3
flask-socketio==5.3.6
python-engineio==4.8.0
python-socketio==5.10.0

# WebRTC
aiortc==1.5.0
aiohttp==3.9.1

# Computer Vision
opencv-python==4.8.1.78
numpy==1.26.2
pillow==10.1.0

# AI and Machine Learning
whisper==1.1.10
langid==1.1.6
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1

# SLAM and Navigation
orbslam2-python==0.3.1  # This is a placeholder, actual package may differ

# Motor Control
pyserial==3.5
RPi.GPIO==0.7.1  # Only for Raspberry Pi

# Text-to-Speech
pyttsx3==2.90
espeak==0.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
websockets==11.0.3 